+++ enlighten/device/LaserControlFeature.py
@@ -167,16 +167,15 @@ class LaserControlFeature:
                     self.configure_laser_power_controls_percent()
 
         # let them turn the laser on/off, nothing else (including watchdog)
-        for w in [ cfu.doubleSpinBox_excitation_nm,
-                   cfu.pushButton_laser_power_dn,
-                   cfu.pushButton_laser_power_up,
-                   cfu.verticalSlider_laser_power,
+        for w in [ cfu.checkBox_laser_watchdog,
+                   cfu.comboBox_laser_power_unit,
+                   cfu.comboBox_laser_tec_mode,
                    cfu.doubleSpinBox_excitation_nm,
                    cfu.doubleSpinBox_laser_power,
+                   cfu.pushButton_laser_power_dn,
+                   cfu.pushButton_laser_power_up,
                    cfu.spinBox_laser_watchdog_sec,
-                   cfu.checkBox_laser_watchdog,
-                   cfu.comboBox_laser_power_unit,
-                   cfu.comboBox_laser_tec_mode ]:
+                   cfu.verticalSlider_laser_power ]:
             w.setEnabled(not self.locked)
 
         self.refresh_laser_buttons()
@@ -515,11 +514,15 @@ class LaserControlFeature:
         if spec is None:
             return
 
-        value = self.ctl.form.ui.doubleSpinBox_excitation_nm.value()
+        cfu = self.ctl.form.ui
+
+        value = cfu.doubleSpinBox_excitation_nm.value()
         log.debug("changing current spectrometer's excitation to %.2f", value)
 
-        spec.settings.eeprom.excitation_nm_float = value
-        self.ctl.eeprom_editor.widget_callback("excitation_nm_float", reset_from_eeprom=True)
+        spec.settings.eeprom.multi_wavelength_calibration.set("excitation_nm_float", value)
+
+        log.debug("calling Settings.update_wavecal")
+        spec.settings.update_wavecal()
 
     # gets called by BatteryFeature when a new battery reading is received
     def battery_callback(self, perc, charging):