+++ CHANGELOG.md
@@ -6,7 +6,9 @@
         - added OEM/Multichannel plugin
         - persist integration time on locked spectrometers to enlighten.ini
         - make Dark lockable
-    - simplified Analysis.PeakFinding plugin
+    - plugins
+        - updated Analysis.PeakFinding
+        - updated Prod.PixelNoise
 - 2024-11-05 4.1.16
     - XS
         - support new startup_laser_tec_setpoint EEPROM field